pyproject.toml
[tool.poetry]
name = "bert-finetuning-catalyst"
version = "0.1.0"
description = "BERT classifier fine-tuning with PyTorch, HuggingFace, and Catalyst"
authors = ["Yury Kashnitsky <yury.kashnitsky@gmail.com>"]
license = "CCBY-4.0"

[tool.poetry.dependencies]
python = ">=3.8,<=3.9.9"
catalyst = "22.4"
numpy = "^1.21.5"
pandas = "^1.3.5"
PyYAML = "^5.3.1"
tensorboard = "^2.8.0"
torch = "^1.10.0"
transformers = "^4.18.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
click = "^8.0.4"
flake8 = "^3.7.9"
isort = "^5.10.1"
mypy = "^0.770"
pre_commit = "^2.18.1"
pylint = "^2.12.2"
pytest = "^5.2"
pytest-cov = "^2.8.1"
pytest-html = "^2.1.0"

[tool.black]
line-length = 119
target-version = ['py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.tox
  | \*venv
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
